[common] 이미지 관련 전반적 수정 - bitmap memory 개선 - image cache 적용 - 이미지 로드 시 로테이션 적용 - 올사 기능 추가 - 채팅방 이미지 안보이는 문제 수정 - 채팅방에서 본인 이미지 추가 - 만남요청이 비정상적으로 동작하던 문제 수정

+++ app/src/main/java/com/ssomcompany/ssomclient/network/api/GetChattingList.java
@@ -13,7 +13,7 @@ import java.util.Locale;
 public class GetChattingList {
 
     public static class Request extends SsomGetRequest {
-        public Request(int roomId) {
+        public Request(long roomId) {
             super(String.format(Locale.getDefault(), API.CHAT_LIST, roomId));
         }
     }
